Introduce custom exceptions and refactor dependency
