+++ openday_scavenger/static/html/404_general.html
@@ -1,0 +1,9 @@
+{% extends "layout.html" %}
+{% block content %}
+<div class="d-flex justify-content-center align-items-center">
+    <div class="col-md-12 text-center">
+        <h1>404</h1>
+        <h2>Unknown Page</h2>
+    </div>
+</div>
+{% endblock %}
